requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymysql==1.1.0
python-dotenv==1.0.0
boto3==1.34.34
bcrypt==4.1.2
PyJWT==2.8.0
Werkzeug==3.0.1
requests==2.31.0
